hugheslegacy
We understand that our area of the law is primarily about relationships and legacies. We work hard to develop and maintain relationships among our clients, staff, and colleagues in the legal profession and, of course, our personal friends and families. We are proud of the legacies we have inherited from attorneys who came before us. Many generous attorneys have educated and mentored us and have honored us by leaving their treasured clients in our hands. We carry those torches into the future with gratitude. We maintain a compelling desire to work with our clients as their relationship needs and goals mature into their own legacies.